requirements/base.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.11.8
cffi==1.14.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
ClickUz==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filer==2.0.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-oauth-toolkit==1.3.3
django-oauth2-provider==0.2.6.1
django-parler==2.2
django-parler-rest==2.1
django-polymorphic==3.0.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
future==0.18.2
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
PyVimeo==1.1.0
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tablib==2.0.0
telebot==0.0.4
tinydb==4.4.0
tuspy==0.2.4
twilio==6.58.0
Twisted==21.2.0
txaio==20.12.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
gunicorn==20.1.0
django-storages
boto
boto3